Refactor 'split_allowed' method and update corresponding tests
The 'split_allowed' method in the player test class has been improved to directly take the split counter as an argument, eliminating the need to manually set it each time. The bet and balance after a player split are also now accurately reflected in the tests. Unnecessary tests related to the old