Refactor deck shuffle mechanism and update network response
This commit removes the test methods for deck shuffling and reshuffles deck only if needed according to a newly introduced 'need_shuffle' boolean field in the player model. It also includes the 'need_shuffle' field in the network response to give the player more information about the game state. Further, it changes the application's host port for unspecified reasons.